+++ examples/admin-sdk-plugin/src/Resources/app/administration/CHANGELOG.md
@@ -1,5 +1,15 @@
 # administration
 
+## 1.0.17
+
+### Patch Changes
+
+- Updated dependencies [243ffd8]
+- Updated dependencies [9b0cc45]
+- Updated dependencies [adb59f8]
+  - @shopware-ag/meteor-component-library@4.10.0
+  - @shopware-ag/meteor-admin-sdk@5.8.5
+
 ## 1.0.16
 
 ### Patch Changes